requirements.txt
nltk==3.3
numpy==1.14.3
panda==0.23.0
pip==18.0
python-dateutil==2.7.3
pytz==2018.4
setuptools==39.1.0
six==1.11.0
tabulate==0.8.2